+++ scru128/__init__.py
@@ -282,6 +282,8 @@ class Scru128Generator:
     @property
     def last_status(self) -> Scru128Generator.Status:
         """
+        Deprecated: use `generate_monotonic()` to guarantee monotonicity.
+
         Returns a `Status` code that indicates the internal state involved in the last
         generation of ID.
 
@@ -289,6 +291,9 @@ class Scru128Generator:
         during the sequential calls to a generation method and this property to avoid
         race conditions.
         """
+        warnings.warn(
+            "use `generate_monotonic()` to guarantee monotonicity", DeprecationWarning
+        )
         return self._last_status
 
     def __iter__(self) -> typing.Iterator[Scru128Id]:
@@ -308,7 +313,7 @@ class Scru128Generator:
 
     class Status(enum.Enum):
         """
-        Status code returned by `last_status` property.
+        Deprecated. Status code returned by `last_status` property.
 
         Attributes:
             NOT_EXECUTED: Indicates that the generator has yet to generate an ID.