+++ tests/test_generator.py
@@ -10,18 +10,18 @@ class TestGenerateCore(unittest.TestCase):
         """Generates increasing IDs even with decreasing or constant timestamp"""
         ts = 0x0123_4567_89AB
         g = Scru128Generator()
-        self.assertEqual(g.last_status, Scru128Generator.Status.NOT_EXECUTED)
+        self.assertEqual(g._last_status, Scru128Generator.Status.NOT_EXECUTED)
 
         prev = g.generate_core(ts)
-        self.assertEqual(g.last_status, Scru128Generator.Status.NEW_TIMESTAMP)
+        self.assertEqual(g._last_status, Scru128Generator.Status.NEW_TIMESTAMP)
         self.assertEqual(prev.timestamp, ts)
 
         for i in range(100_000):
             curr = g.generate_core(ts - min(9_998, i))
             self.assertTrue(
-                g.last_status == Scru128Generator.Status.COUNTER_LO_INC
-                or g.last_status == Scru128Generator.Status.COUNTER_HI_INC
-                or g.last_status == Scru128Generator.Status.TIMESTAMP_INC
+                g._last_status == Scru128Generator.Status.COUNTER_LO_INC
+                or g._last_status == Scru128Generator.Status.COUNTER_HI_INC
+                or g._last_status == Scru128Generator.Status.TIMESTAMP_INC
             )
             self.assertLess(prev, curr)
             prev = curr
@@ -31,14 +31,14 @@ class TestGenerateCore(unittest.TestCase):
         """Breaks increasing order of IDs if timestamp moves backward a lot"""
         ts = 0x0123_4567_89AB
         g = Scru128Generator()
-        self.assertEqual(g.last_status, Scru128Generator.Status.NOT_EXECUTED)
+        self.assertEqual(g._last_status, Scru128Generator.Status.NOT_EXECUTED)
 
         prev = g.generate_core(ts)
-        self.assertEqual(g.last_status, Scru128Generator.Status.NEW_TIMESTAMP)
+        self.assertEqual(g._last_status, Scru128Generator.Status.NEW_TIMESTAMP)
         self.assertEqual(prev.timestamp, ts)
 
         curr = g.generate_core(ts - 10_000)
-        self.assertEqual(g.last_status, Scru128Generator.Status.CLOCK_ROLLBACK)
+        self.assertEqual(g._last_status, Scru128Generator.Status.CLOCK_ROLLBACK)
         self.assertGreater(prev, curr)
         self.assertEqual(curr.timestamp, ts - 10_000)
 
@@ -57,20 +57,20 @@ class TestGenerateCoreMonotonic(unittest.TestCase):
         """Generates increasing IDs even with decreasing or constant timestamp"""
         ts = 0x0123_4567_89AB
         g = Scru128Generator()
-        self.assertEqual(g.last_status, Scru128Generator.Status.NOT_EXECUTED)
+        self.assertEqual(g._last_status, Scru128Generator.Status.NOT_EXECUTED)
 
         prev = g.generate_core_monotonic(ts)
         assert prev is not None
-        self.assertEqual(g.last_status, Scru128Generator.Status.NEW_TIMESTAMP)
+        self.assertEqual(g._last_status, Scru128Generator.Status.NEW_TIMESTAMP)
         self.assertEqual(prev.timestamp, ts)
 
         for i in range(100_000):
             curr = g.generate_core_monotonic(ts - min(9_998, i))
             assert curr is not None
             self.assertTrue(
-                g.last_status == Scru128Generator.Status.COUNTER_LO_INC
-                or g.last_status == Scru128Generator.Status.COUNTER_HI_INC
-                or g.last_status == Scru128Generator.Status.TIMESTAMP_INC
+                g._last_status == Scru128Generator.Status.COUNTER_LO_INC
+                or g._last_status == Scru128Generator.Status.COUNTER_HI_INC
+                or g._last_status == Scru128Generator.Status.TIMESTAMP_INC
             )
             self.assertLess(prev, curr)
             prev = curr
@@ -80,16 +80,16 @@ class TestGenerateCoreMonotonic(unittest.TestCase):
         """Returns None if timestamp moves backward a lot"""
         ts = 0x0123_4567_89AB
         g = Scru128Generator()
-        self.assertEqual(g.last_status, Scru128Generator.Status.NOT_EXECUTED)
+        self.assertEqual(g._last_status, Scru128Generator.Status.NOT_EXECUTED)
 
         prev = g.generate_core_monotonic(ts)
         assert prev is not None
-        self.assertEqual(g.last_status, Scru128Generator.Status.NEW_TIMESTAMP)
+        self.assertEqual(g._last_status, Scru128Generator.Status.NEW_TIMESTAMP)
         self.assertEqual(prev.timestamp, ts)
 
         curr = g.generate_core_monotonic(ts - 10_000)
         assert curr is None
-        self.assertEqual(g.last_status, Scru128Generator.Status.NEW_TIMESTAMP)
+        self.assertEqual(g._last_status, Scru128Generator.Status.NEW_TIMESTAMP)
 
         curr = g.generate_core_monotonic(ts - 10_001)
         assert curr is None